% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslr-package.R
\docType{package}
\name{tslr-package}
\alias{tslr}
\alias{tslr-package}
\title{tslr: R Bindings for Tensor Separation Learning}
\description{
R bindings for TSL (Tensor Separation Learning), a glass-box regression model implemented in Rust. Provides a simple fit/predict interface backed by the high-performance 'tsl' Rust crate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jyliuu/TSL}
  \item Report bugs at \url{https://github.com/jyliuu/TSL/issues}
}

}
\author{
\strong{Maintainer}: Jinyang Liu \email{jinyang.liu.jl@gmail.com} [copyright holder]

}
\keyword{internal}
